fix(ollama,api): resolve server errors from invalid option parameters
- Issue: Ollama rejecting API options with "invalid option provided" errors
- Root cause: JSON serialization using camelCase instead of snake_case
- Solution: Added explicit JsonPropertyName attributes for correct API format
- Fixed: num_ctx parameter now works with user's context window setting
- Improved: Clean JSON serialization ignoring null values
- Result: Zero server errors, UI progress bar and status messages functional

Files modified:
- Services/OllamaTypes.cs: Added JsonPropertyName attributes for snake_case
- Services/OllamaHttpService.cs: Added null value ignoring in JSON serialization
- Services/OllamaService.cs: Restored numCtx option with correct formatting
