+++ Docs/IMPLEMENTATION_PLAN.md
@@ -109,16 +109,18 @@ This document outlines the implementation plan to fix all critical issues identi
 ### 2.2 God Class Refactoring
 **Files Affected:** `Services/OllamaService.cs`
 
-- [ ] **Split OllamaService into focused services**
-  - Create `OllamaHttpService` for HTTP operations
-  - Create `OllamaProcessService` for process management
-  - Create `OllamaPerformanceService` for metrics
-  - Create `OllamaFileService` for file operations
+- [x] **Split OllamaService into focused services**
+  - Created `OllamaHttpService` for HTTP operations
+  - Created `OllamaProcessService` for process management
+  - Created `OllamaPerformanceService` for metrics
+  - Created `OllamaFileService` for file operations
+  - Created `OllamaTypes.cs` for shared types
 
-- [ ] **Implement proper separation of concerns**
-  - Each service should have single responsibility
+- [x] **Implement proper separation of concerns**
+  - Each service has single responsibility
   - Use dependency injection between new services
   - Implement proper interface segregation
+  - Main OllamaService now acts as orchestrator
 
 ### 2.3 Exception Handling Improvements
 **Files Affected:** Multiple service files
@@ -311,4 +313,4 @@ This document outlines the implementation plan to fix all critical issues identi
 ---
 
 *Last Updated: September 17, 2025*
-*Status: Phase 2.1 (Circular Dependencies & DI Patterns) - ✅ COMPLETED. Ready for Phase 2.2 (God Class Refactoring)*
+*Status: Phase 2.2 (God Class Refactoring) - ✅ COMPLETED. Ready for Phase 2.3 (Exception Handling Improvements)*
